+++ src/frontend/contexts/index.js
@@ -1,0 +1,4 @@
+export { useAuth } from "./auth-context";
+export { useCart } from "./cart-context";
+export { useData } from "./data-context";
+export { useProducts } from "./products-context";
