+++ src/frontend/pages/Products/components/ProductsListing.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useProducts } from "../../../contexts/products-context";
+import { useProducts } from "../../../contexts/index";
 import { Card } from "../../../index.js";
 
 export function ProductsListing() {
