+++ src/frontend/pages/index.js
@@ -1,4 +1,5 @@
 export { Products } from "./Products/Products.jsx";
 export { Home } from "./Home/Home.jsx";
 export { Login } from "./Authentication/Login.jsx";
+export { Login } from "./Authentication/Logout.jsx";
 export { Cart } from "./Cart/Cart.jsx";
